Review suggestion: remove redundant state
