+++ assets/js/dashboard/components/query-context-hoc.js
@@ -2,7 +2,6 @@ import React, { useState, useEffect} from "react"
 import * as api from '../api'
 import { useMountedEffect } from '../custom-hooks'
 import { parseQuery } from "../query"
-import { getFiltersByKeyPrefix } from '../util/filters'
 
 // A Higher-Order component that tracks `query` state, and additional context
 // related to it, such as:
@@ -10,11 +9,6 @@ import { getFiltersByKeyPrefix } from '../util/filters'
 // * `importedDataInView` - simple state with a `false` default. An
 //   `updateImportedDataInView` prop will be passed into the WrappedComponent
 //   and allows changing that according to responses from the API.
-
-// * `revenueAvailable` - keeps track of whether the current query includes a
-//   non-empty goal filterset containing a single, or multiple revenue goals
-//   with the same currency. Can be used to decide whether to render revenue
-//   metrics in a dashboard report or not.
 
 // * `lastLoadTimestamp` - used for displaying a tooltip with time passed since
 //   the last update in realtime components.
@@ -25,7 +19,6 @@ export default function withQueryContext(WrappedComponent) {
     
     const [query, setQuery] = useState(parseQuery(location.search, site))
     const [importedDataInView, setImportedDataInView] = useState(false)
-    const [revenueAvailable, setRevenueAvailable] = useState(false)
     const [lastLoadTimestamp, setLastLoadTimestamp] = useState(new Date())
     
     const updateLastLoadTimestamp = () => { setLastLoadTimestamp(new Date()) }
@@ -38,22 +31,6 @@ export default function withQueryContext(WrappedComponent) {
       }
     }, [])
 
-    useEffect(() => {
-      const revenueGoalsInFilter = site.revenueGoals.filter((rg) => {
-        const goalFilters = getFiltersByKeyPrefix(query, "goal")
-        
-        return goalFilters.some(([_op, _key, clauses]) => {
-          return clauses.includes(rg.event_name)
-        })
-      })
-
-      const singleCurrency = revenueGoalsInFilter.every((rg) => {
-        return rg.currency === revenueGoalsInFilter[0].currency
-      })
-
-      setRevenueAvailable(revenueGoalsInFilter.length > 0 && singleCurrency)
-    }, [query])
-
     useMountedEffect(() => {
       api.cancelAll()
       setQuery(parseQuery(location.search, site))
@@ -64,7 +41,6 @@ export default function withQueryContext(WrappedComponent) {
       <WrappedComponent
         {...props}
         query={query}
-        revenueAvailable={revenueAvailable}
         importedDataInView={importedDataInView}
         updateImportedDataInView={setImportedDataInView}
         lastLoadTimestamp={lastLoadTimestamp}
